+++ sharepasswordAzure/Options/ConsoleAuditLoggingOptions.cs
@@ -1,0 +1,9 @@
+namespace SharePassword.Options;
+
+public class ConsoleAuditLoggingOptions
+{
+    public const string SectionName = "ConsoleAuditLogging";
+
+    public bool Enabled { get; set; } = false;
+    public string Level { get; set; } = "INFO";
+}